requirements.txt
#kivy
requests
flask
numpy
google-generativeai
deepl
dotenv
argon2-cffi
langdetect
PyJwt
flask-cors